requirements_GIP2018.txt
absl-py==1.4.0
astor==0.8.1
backcall==0.2.0
cached-property==1.5.2
colorama==0.4.5
cycler==0.11.0
Cython==3.0.3
dataclasses==0.8
decorator==4.4.2
gast==0.5.4
google-pasta==0.2.0
grpcio==1.48.2
h5py==2.10.0
imageio==2.15.0
importlib-metadata==4.8.3
ipython==7.8.0
ipython-genutils==0.2.0
jedi==0.19.1
Keras==2.2.0
Keras-Applications==1.0.2
Keras-Preprocessing==1.0.1
kiwisolver==1.3.1
Markdown==3.3.7
matplotlib==3.3.4
networkx==2.5.1
numpy==1.19.5
opencv-python==4.1.1.26
parso==0.8.3
pickleshare==0.7.5
Pillow==8.4.0
prompt-toolkit==2.0.10
protobuf==3.19.6
pycocotools-win==2.0
Pygments==2.14.0
pyparsing==3.1.1
python-dateutil==2.8.2
PyWavelets==1.1.1
PyYAML==6.0.1
questionary==1.5.1
scikit-image==0.15.0
scipy==1.3.1
six==1.16.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
tqdm==4.43.0
traitlets==4.3.3
typing_extensions==4.1.1
wcwidth==0.2.8
Werkzeug==2.0.3
wrapt==1.15.0
zipp==3.6.0
